Move basic_evm_add_test into preinstalled_contracts.rs
